// shared/enhancements.generated.ts
export const STOCK_BINDINGS = [
  {
    name: 'Singularity Matrix',
    datasheet: "C'tan Shard of the Deceiver",
    cost: 45,
    detachment: 'stock-detachment:necrons:pantheonofwoe',
  },
  {
    name: 'Quantum Goad',
    datasheet: "C'tan Shard of the Nightbringer",
    cost: 45,
    detachment: 'stock-detachment:necrons:pantheonofwoe',
  },
  {
    name: 'Animus Damper',
    datasheet: "C'tan Shard of the Void Dragon",
    cost: 35,
    detachment: 'stock-detachment:necrons:pantheonofwoe',
  },
  {
    name: 'Relativistic Tether',
    datasheet: "Transcendent C'tan",
    cost: 40,
    detachment: 'stock-detachment:necrons:pantheonofwoe',
  },
]